This is my initial commit.
The full game is completed in this commit, however, I would like to leverage taipy to add a more visual component to the program.
The idea using taipy is that the user can see their selection in real-time, be fed the results, and continue the game
all through the UI. If I am not able to get around to the creation of this solution leveraging taipy, this game is meant
to be played by taking images of the player's hand and determining whether their selection is rock, paper, or scissors.
The computer's choice is randomized, and compared to the player's. The first to 5 points wins!
The model was trained using Teachable Machine, and the images that were used were from a post on Kaggle. There is a ReadMe in the folder containing
the training images that highlight the source and the description of the materials.